Fixed a printf format error when compiling in 64bit mode.
git-svn-id: https://shellinabox.googlecode.com/svn/trunk@30 0da03de8-d603-11dd-86c2-0f8696b7b6f9

+++ libhttp/httpconnection.c
@@ -1277,11 +1277,11 @@ void httpSendReply(struct HttpConnection *http, int code,
                                 "HTTP/1.1 %d %s\r\n"
                                 "%s"
                                 "Content-Type: text/html; charset=utf-8\r\n"
-                                "Content-Length: %d\r\n"
+                                "Content-Length: %ld\r\n"
                                 "\r\n",
                                 code, msg,
                                 code != 200 ? "Connection: close\r\n" : "",
-                                strlen(body));
+                                (long)strlen(body));
   }
   int isHead     = !strcmp(http->method, "HEAD");
   if (!isHead) {
